+ [ Floating point presision ]
main.c:
	- changed fahr and calcius from ints to floats
	- rearanged calculation and converted numbers to floats
	- changed printf from subsiting as ints to floats and changed formatting
	  to align numbers

+++ 1/2/main.c
@@ -4,7 +4,7 @@ main()
 {
 	/* farh holds the current temp to be converted, celsius holds the
 	   converted temperature*/
-	int farh, celsius;
+	float farh, celsius;
 
 	/* range of conversions for farh*/
 	int start, stop, step;
@@ -15,13 +15,13 @@ main()
 
 	printf("°F\t°C\n");
 	/* farh acts as counter */
-	farh = start;
+	farh = (float) start;
 	/* loop till counter (farh) is bigger than stop, inclusive range */
 	while (farh <= stop)
 	{ 
 		/* convert farh to celcius */
-		celsius =  ((farh - 32) * 5) / 9;
-		printf("%d\t%d\n",farh, celsius);
+		celsius = (farh - 32) * (5.0 / 9.0);
+		printf("%4.0f\t%6.1f\n", farh, celsius);
 		/* increemnt counter (farh) */
 		farh = farh + step;
 	}
